Create Ronnie 2k
By this we have known love, that He laid down His life for us; and we ought to lay down our lives for the brothers. But whoever has the world's goods, and sees his brother in need and closes his heart against him, how does the love of God abide in him? Little children, let us not love with word or with tongue, but in deed and truth.